requirements.txt
flask==2.3.3
aiohttp==3.9.1
openai==1.5.0
python-dotenv==1.0.0